refactor launch logic
