+++ tests/unit/commands/test_launch.py
@@ -1,18 +1,15 @@
 from pathlib import Path
 from typing import List, Optional
-from unittest.mock import MagicMock
+from unittest.mock import MagicMock, PropertyMock
 
 from databricks_cli.sdk import JobsService
 from pytest_mock import MockFixture
 
+from dbx.api.client_provider import DatabricksClientProvider
 from dbx.api.config_reader import ConfigReader
-from dbx.commands.launch import (
-    _cancel_run,
-    _load_dbx_file,
-    _trace_run,
-)
+from dbx.api.launch.tracer import RunTracer
 from dbx.utils.json import JsonUtils
-from tests.unit.conftest import extract_function_name, invoke_cli_runner
+from tests.unit.conftest import invoke_cli_runner
 
 
 def deploy_and_get_job_name(deploy_args: Optional[List[str]] = None) -> str:
@@ -146,7 +143,7 @@ def test_launch_run_submit(
     deployment_result = Path(".dbx/deployment-result.json")
     _chosen_job = deploy_and_get_job_name(["--files-only", "--write-specs-to-file", deployment_result])
     mocked_result = JsonUtils.read(deployment_result)
-    mocker.patch(extract_function_name(_load_dbx_file), MagicMock(return_value=mocked_result))
+    mocker.patch("dbx.api.launch.runners.load_dbx_file", MagicMock(return_value=mocked_result))
     launch_result = invoke_cli_runner(["launch", "--job", _chosen_job] + ["--as-run-submit"])
     assert launch_result.exit_code == 0
 
@@ -176,12 +173,30 @@ def test_launch_with_output(
     assert launch_result.exit_code == 0
 
 
-def test_launch_with_run_now_params(
-    mocker: MockFixture, temp_project: Path, mlflow_file_uploader, mock_dbx_file_upload, mock_api_v2_client
+def test_launch_with_run_now_v21_params(
+    mocker: MockFixture, temp_project: Path, mlflow_file_uploader, mock_dbx_file_upload
 ):
+    client_mock = MagicMock()
+    p = PropertyMock(return_value="2.1")
+    type(client_mock).jobs_api_version = p
+    mocker.patch.object(DatabricksClientProvider, "get_v2_client", lambda: client_mock)
     _chosen_job = deploy_and_get_job_name()
     prepare_job_service_mock(mocker, _chosen_job)
-    launch_result = invoke_cli_runner(["launch", "--job", _chosen_job, "--parameters", '{"python_params":["a", "b"]}'])
+    launch_result = invoke_cli_runner(
+        ["launch", "--job", _chosen_job, "--parameters", '{"python_named_params":{"a":1}}']
+    )
+    assert launch_result.exit_code == 0
+
+
+def test_launch_with_run_now_v20_params(
+    mocker: MockFixture, temp_project: Path, mlflow_file_uploader, mock_dbx_file_upload
+):
+    client_mock = MagicMock()
+    type(client_mock).jobs_api_version = PropertyMock(return_value="2.0")
+    mocker.patch.object(DatabricksClientProvider, "get_v2_client", lambda: client_mock)
+    _chosen_job = deploy_and_get_job_name()
+    prepare_job_service_mock(mocker, _chosen_job)
+    launch_result = invoke_cli_runner(["launch", "--job", _chosen_job, "--parameters", '{"python_params":[1,2]}'])
     assert launch_result.exit_code == 0
 
 
@@ -224,10 +239,9 @@ def test_launch_with_trace_and_kill_on_sigterm_with_interruption(
 ):
     _chosen_job = deploy_and_get_job_name(["--tags", "soup=beautiful"])
     prepare_job_service_mock(mocker, _chosen_job)
-    mocker.patch(extract_function_name(_trace_run), MagicMock(side_effect=[KeyboardInterrupt("stopped!")])),
-    cancel_run_mock = mocker.patch(extract_function_name(_cancel_run))
+    _tracer = mocker.patch.object(RunTracer, "start", return_value=("SUCCESS", {}))
     launch_result = invoke_cli_runner(
         ["launch", "--job", _chosen_job] + ["--tags", "soup=beautiful", "--trace", "--kill-on-sigterm"]
     )
     assert launch_result.exit_code == 0
-    cancel_run_mock.assert_called_once()
+    _tracer.assert_called_once()
